converted to maven project
